refact: 책임 분리에 따른 테스트코드 분리 board는 순수 도메인으로 변경. 각종 기능은 별도 클래스로 분리
